requirements.txt
# requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
supabase==2.28.3
pymongo==4.16.0
groq==0.37.1
python-dotenv==1.2.1
python-multipart==0.0.2
pydantic==2.12.5